# webapp/docs/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/manager/reservations" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReservationResponseDTO"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          }
        },
        "summary" : "Get All Reservations",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "ids",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Reservation deleted successfully"
          },
          "400" : {
            "description" : "Bad Request: Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Reservation with specified ID not found for the current manager"
          }
        },
        "summary" : "Delete Reservation",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReservationRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReservationResponseDTO"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request: Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Event, user or seat not found"
          },
          "409" : {
            "description" : "Conflict: Seat already reserved or event booking closed"
          }
        },
        "summary" : "Create Reservations",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      }
    },
    "/api/manager/reservations/block" : {
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BlockSeatsRequestDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Seats blocked successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReservationResponseDTO"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request: Invalid input"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Event or seat not found"
          },
          "409" : {
            "description" : "Conflict: Seat already blocked"
          }
        },
        "summary" : "Block Seats",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      }
    },
    "/api/manager/reservations/event/{id}" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ReservationResponseDTO"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Event with specified ID not found for the current manager"
          }
        },
        "summary" : "Get Reservations By Event Id",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      }
    },
    "/api/manager/reservations/export/{eventId}/csv" : {
      "get" : {
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "CSV export of reservations for a specific event",
            "content" : {
              "text/csv" : { }
            }
          },
          "403" : {
            "description" : "Forbidden - User not authorized"
          },
          "404" : {
            "description" : "Not Found - Event not found"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "Internal Server Error during CSV export"
          }
        },
        "summary" : "Export Reservations To Csv",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      }
    },
    "/api/manager/reservations/export/{eventId}/pdf" : {
      "get" : {
        "parameters" : [ {
          "name" : "eventId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "PDF export of reservations for a specific event",
            "content" : {
              "application/pdf" : { }
            }
          },
          "403" : {
            "description" : "Forbidden - User not authorized"
          },
          "404" : {
            "description" : "Not Found - Event not found"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "500" : {
            "description" : "Internal Server Error during PDF export"
          }
        },
        "summary" : "Export Reservations To Pdf",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      }
    },
    "/api/manager/reservations/{id}" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only MANAGER or ADMIN roles can access this resource"
          },
          "404" : {
            "description" : "Not Found: Reservation with specified ID not found for the current manager"
          }
        },
        "summary" : "Get Reservation By Id",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "MANAGER", "ADMIN" ]
        } ]
      }
    },
...
    "/api/user/reservations/{id}" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserReservationResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden: Only authenticated users can access this resource"
          },
          "404" : {
            "description" : "Not Found: Reservation with specified ID not found for the current user"
          }
        },
        "summary" : "Get My Reservation By Id",
        "tags" : [ "Reservation Resource" ],
        "security" : [ {
          "SecurityScheme" : [ "USER", "MANAGER", "ADMIN" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BlockSeatsRequestDTO" : {
        "type" : "object",
        "properties" : {
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "seatIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
...
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
...
      "ReservationRequestDTO" : {
        "type" : "object",
        "required" : [ "eventId", "userId", "seatIds" ],
        "properties" : {
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "seatIds" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "deductAllowance" : {
            "type" : "boolean"
          }
        }
      },
      "ReservationResponseDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserDTO"
          },
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "seat" : {
            "$ref" : "#/components/schemas/SeatDTO"
          },
          "reservationDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "status" : {
            "$ref" : "#/components/schemas/ReservationStatus"
          }
        }
      },
      "ReservationStatus" : {
        "type" : "string",
        "enum" : [ "RESERVED", "BLOCKED" ]
      },
      "SeatDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "seatNumber" : {
            "type" : "string"
          },
          "seatRow" : {
            "type" : "string"
          },
          "locationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "xCoordinate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "yCoordinate" : {
            "type" : "integer",
            "format" : "int32"
          },
          "entrance" : {
            "type" : "string"
          }
        }
      },
...
      "UserDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "emailVerified" : {
            "type" : "boolean"
          },
          "emailVerificationSent" : {
            "type" : "boolean"
          },
          "roles" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string"
            }
          },
          "tags" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "UserReservationResponseDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "seat" : {
            "$ref" : "#/components/schemas/SeatDTO"
          },
          "reservationDateTime" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT",
        "description" : "Authentication"
      }
    }
  },
...
  "info" : {
    "title" : "Seat Reservation API",
    "version" : "0.0.7",
    "description" : "The official REST API for Seat Reservation.",
    "contact" : {
      "name" : "FelixHertweck GitHub ",
      "url" : "https://github.com/FelixHertweck"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}